+++ src/userPermissionsApi.js
@@ -17,9 +17,8 @@ class UserPermissionsApi {
 
   async authorizeUser(userId, resourceUri, permission) {
     // verify required parameter 'userId' is not null or undefined
-    let tokenUserId = userId;
     if (userId === null || userId === undefined) {
-      tokenUserId = await getFallbackUser(this.client);
+      throw new ArgumentRequiredError('userId', 'Required parameter userId was null or undefined when calling authorizeUser.');
     }
     // verify required parameter 'resourceUri' is not null or undefined
     if (resourceUri === null || resourceUri === undefined) {
@@ -29,13 +28,13 @@ class UserPermissionsApi {
     if (permission === null || permission === undefined) {
       throw new ArgumentRequiredError('permission', 'Required parameter permission was null or undefined when calling authorizeUser.');
     }
-    const url = `/v1/users/${encodeURIComponent(String(tokenUserId))}/resources/${encodeURIComponent(String(resourceUri))}/permissions/${encodeURIComponent(String(permission))}`;
+    const url = `/v1/users/${encodeURIComponent(String(userId))}/resources/${encodeURIComponent(String(resourceUri))}/permissions/${encodeURIComponent(String(permission))}`;
     try {
       const response = await this.client.get(url);
       return response;
     } catch (error) {
       if (error.status === 404) {
-        throw new UnauthorizedError(tokenUserId, resourceUri, permission);
+        throw new UnauthorizedError(userId, resourceUri, permission);
       }
       throw error;
     }
@@ -43,77 +42,54 @@ class UserPermissionsApi {
 
   async disableUserToken(userId, tokenId) {
     // verify required parameter 'userId' is not null or undefined
-    let tokenUserId = userId;
     if (userId === null || userId === undefined) {
-      tokenUserId = await getFallbackUser(this.client);
+      throw new ArgumentRequiredError('userId', 'Required parameter userId was null or undefined when calling disableUserToken.');
     }
     // verify required parameter 'tokenId' is not null or undefined
     if (tokenId === null || tokenId === undefined) {
       throw new ArgumentRequiredError('tokenId', 'Required parameter tokenId was null or undefined when calling disableUserToken.');
     }
-    const url = `/v1/users/${encodeURIComponent(String(tokenUserId))}/tokens/${encodeURIComponent(String(tokenId))}`;
+    const url = `/v1/users/${encodeURIComponent(String(userId))}/tokens/${encodeURIComponent(String(tokenId))}`;
     const response = await this.client.delete(url);
     return response;
   }
 
   async getUserPermissionsForResource(userId, resourceUri) {
     // verify required parameter 'userId' is not null or undefined
-    let tokenUserId = userId;
     if (userId === null || userId === undefined) {
-      tokenUserId = await getFallbackUser(this.client);
+      throw new ArgumentRequiredError('userId', 'Required parameter userId was null or undefined when calling getUserPermissionsForResource.');
     }
     // verify required parameter 'resourceUri' is not null or undefined
     if (resourceUri === null || resourceUri === undefined) {
       throw new ArgumentRequiredError('resourceUri', 'Required parameter resourceUri was null or undefined when calling getUserPermissionsForResource.');
     }
 
-    const url = `/v1/users/${encodeURIComponent(String(tokenUserId))}/resources/${encodeURIComponent(String(resourceUri))}/permissions`;
+    const url = `/v1/users/${encodeURIComponent(String(userId))}/resources/${encodeURIComponent(String(resourceUri))}/permissions`;
     const response = await this.client.get(url);
     return response;
   }
 
   async getUserRolesForResource(userId, resourceUri) {
-    let tokenUserId = userId;
     if (userId === null || userId === undefined) {
-      tokenUserId = await getFallbackUser(this.client);
+      throw new ArgumentRequiredError('userId', 'Required parameter userId was null or undefined when calling getUserRolesForResource.');
     }
     // verify required parameter 'resourceUri' is not null or undefined
     if (resourceUri === null || resourceUri === undefined) {
       throw new ArgumentRequiredError('resourceUri', 'Required parameter resourceUri was null or undefined when calling getUserRolesForResource.');
     }
 
-    const url = `/v1/users/${encodeURIComponent(String(tokenUserId))}/resources/${encodeURIComponent(String(resourceUri))}/roles`;
+    const url = `/v1/users/${encodeURIComponent(String(userId))}/resources/${encodeURIComponent(String(resourceUri))}/roles`;
     const response = await this.client.get(url);
     return response;
   }
 
   async getUserResources(userId, resourceUri, limit, cursor, permission, collectionConfiguration) {
     // verify required parameter 'userId' is not null or undefined
-    let tokenUserId = userId;
     if (userId === null || userId === undefined) {
-      tokenUserId = await getFallbackUser(this.client);
+      throw new ArgumentRequiredError('userId', 'Required parameter userId was null or undefined when calling getUserResources.');
     }
 
-    // If just checking the top level or the collectionConfiguration isn't set then assume we need to check for explicit top level.
-    // * Otherwise we can recurse down using the collectionConfiguration type.
-    if (!collectionConfiguration || collectionConfiguration === 'TOP_LEVEL_ONLY') {
-      try {
-        await this.authorizeUser(userId, resourceUri, permission);
-        return {
-          status: 200,
-          headers: {},
-          data: {
-            userId: tokenUserId,
-            accessToAllSubResources: true,
-            resources: null
-          }
-        };
-      } catch (error) {
-        // If the user doesn't have permission to everything, then use the query API.
-      }
-    }
-
-    const url = new URL(`${this.client.baseUrl}/v1/users/${encodeURIComponent(String(tokenUserId))}/resources`);
+    const url = new URL(`${this.client.baseUrl}/v1/users/${encodeURIComponent(String(userId))}/resources`);
     const qs = { resourceUri };
     if (limit) { qs.limit = limit; }
     if (cursor) { qs.cursor = cursor; }