+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change log
 This is the changelog for [Authress SDK](readme.md).
 
+## 3.0 ##
+* [Breaking] UserId is now required in all `userPermissions` apis. This improves **Security By Default** requiring explicit check on who the user is.
+* [Breaking] Removal of property `accessToAllSubResources`.
+
 ## 2.3 ##
 * Require minimum Node version to be 16.
 * Improve support for collectionConfiguration to skip `accessToAllSubResources` check when using `INCLUDE_NESTED` in the query.